Refactor read_config function to be in one place
