+++ src/cli_shared/mod.rs
@@ -4,6 +4,9 @@
 pub mod cli;
 pub mod logger;
 
+use crate::cli_shared::cli::{find_config_path, Config, ConfigPath};
+use crate::networks::NetworkChain;
+use crate::utils::io::{read_file_to_string, read_toml};
 use std::path::PathBuf;
 
 #[cfg(feature = "mimalloc")]
@@ -12,8 +15,27 @@ pub use mimalloc;
 pub use tikv_jemallocator;
 
 /// Gets chain data directory
-pub fn chain_path(config: &crate::cli_shared::cli::Config) -> PathBuf {
+pub fn chain_path(config: &Config) -> PathBuf {
     PathBuf::from(&config.client.data_dir).join(config.chain.to_string())
+}
+
+pub fn read_config(
+    config_path_opt: &Option<String>,
+    chain_opt: &Option<NetworkChain>,
+) -> anyhow::Result<(Option<ConfigPath>, Config)> {
+    let (path, mut config) = match find_config_path(config_path_opt) {
+        Some(path) => {
+            // Read from config file
+            let toml = read_file_to_string(path.to_path_buf())?;
+            // Parse and return the configuration file
+            (Some(path), read_toml(&toml)?)
+        }
+        None => (None, Config::default()),
+    };
+    if let Some(chain) = chain_opt {
+        config.chain = chain.clone();
+    }
+    Ok((path, config))
 }
 
 pub mod snapshot;
